Backend - Billing Automation/requirements.txt
alembic==1.7.7
amqp==5.1.1
anyio==3.5.0
asgiref==3.4.1
bcrypt==3.2.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
cffi==1.15.0
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
contextlib2==21.6.0
contextvars==2.4
dataclasses==0.8
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
eventlet==0.33.0
fastapi==0.75.1
fastapi-utils==0.2.1
greenlet==1.1.2
h11==0.13.0
httptools==0.3.0
idna==3.3
immutables==0.17
importlib-metadata==4.8.3
importlib-resources==5.4.0
kombu==5.1.0
Mako==1.1.6
MarkupSafe==2.0.1
passlib==1.7.4
prompt-toolkit==3.0.29
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
python-dotenv==0.20.0
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
rsa==4.8
sendgrid==6.9.7
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
starkbank-ecdsa==2.0.3
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.16.0
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==9.1
zipp==3.6.0
